+++ src/quacc/recipes/newtonnet/ts.py
@@ -7,7 +7,7 @@ from monty.dev import requires
 
 from quacc import SETTINGS, job
 from quacc.recipes.newtonnet.core import _add_stdev_and_hess, freq_job, relax_job
-from quacc.runners.calc import run_ase_opt
+from quacc.runners.ase import run_opt
 from quacc.schemas.ase import summarize_opt_run
 from quacc.utils.dicts import merge_dicts
 
@@ -49,9 +49,8 @@ def ts_job(
     use_custom_hessian: bool = False,
     run_freq: bool = True,
     freq_job_kwargs: dict[str, Any] | None = None,
-    calc_swaps: dict[str, Any] | None = None,
-    opt_swaps: dict[str, Any] | None = None,
-    copy_files: list[str] | None = None,
+    opt_params: dict[str, Any] | None = None,
+    **kwargs,
 ) -> TSSchema:
     """
     Perform a transition state (TS) job using the given atoms object.
@@ -66,23 +65,10 @@ def ts_job(
         Whether to run the frequency job.
     freq_job_kwargs
         Keyword arguments to use for the [quacc.recipes.newtonnet.ts.freq_job][]
-    calc_swaps
-        Dictionary of custom kwargs for the NewtonNet calculator. Set a value to
-        `None` to remove a pre-existing key entirely. For a list of available
-        keys, refer to the `newtonnet.utils.ase_interface.MLAseCalculator` calculator.
-
-        !!! Info "Calculator defaults"
-
-            ```python
-            {
-                "model_path": SETTINGS.NEWTONNET_MODEL_PATH,
-                "settings_path": SETTINGS.NEWTONNET_CONFIG_PATH,
-            }
-            ```
-    opt_swaps
+    opt_params
         Dictionary of custom kwargs for the optimization process. Set a value
         to `None` to remove a pre-existing key entirely. For a list of available
-        keys, refer to [quacc.runners.calc.run_ase_opt][].
+        keys, refer to [quacc.runners.ase.run_opt][].
 
         !!! Info "Optimizer defaults"
 
@@ -96,8 +82,19 @@ def ts_job(
                 else {"order": 1},
             }
             ```
-    copy_files
-        Files to copy to the runtime directory.
+    **kwargs
+        Dictionary of custom kwargs for the NewtonNet calculator. Set a value to
+        `None` to remove a pre-existing key entirely. For a list of available
+        keys, refer to the `newtonnet.utils.ase_interface.MLAseCalculator` calculator.
+
+        !!! Info "Calculator defaults"
+
+            ```python
+            {
+                "model_path": SETTINGS.NEWTONNET_MODEL_PATH,
+                "settings_path": SETTINGS.NEWTONNET_CONFIG_PATH,
+            }
+            ```
 
     Returns
     -------
@@ -119,8 +116,8 @@ def ts_job(
         else {"order": 1},
     }
 
-    flags = merge_dicts(defaults, calc_swaps)
-    opt_flags = merge_dicts(opt_defaults, opt_swaps)
+    flags = merge_dicts(defaults, kwargs)
+    opt_flags = merge_dicts(opt_defaults, opt_params)
 
     atoms.calc = NewtonNet(**flags)
 
@@ -131,7 +128,7 @@ def ts_job(
     atoms.calc = ml_calculator
 
     # Run the TS optimization
-    dyn = run_ase_opt(atoms, copy_files=copy_files, **opt_flags)
+    dyn = run_opt(atoms, **opt_flags)
     opt_ts_summary = _add_stdev_and_hess(
         summarize_opt_run(dyn, additional_fields={"name": "NewtonNet TS"})
     )
@@ -155,9 +152,8 @@ def irc_job(
     direction: Literal["forward", "reverse"] = "forward",
     run_freq: bool = True,
     freq_job_kwargs: dict[str, Any] | None = None,
-    calc_swaps: dict[str, Any] | None = None,
-    opt_swaps: dict[str, Any] | None = None,
-    copy_files: list[str] | None = None,
+    opt_params: dict[str, Any] | None = None,
+    **kwargs,
 ) -> IRCSchema:
     """
     Perform an intrinsic reaction coordinate (IRC) job using the given atoms object.
@@ -172,23 +168,10 @@ def irc_job(
         Whether to run the frequency analysis.
     freq_job_kwargs
         Keyword arguments to use for the [quacc.recipes.newtonnet.ts.freq_job][]
-    calc_swaps
-        Dictionary of custom kwargs for the NewtonNet calculator. Set a value to
-        `None` to remove a pre-existing key entirely. For a list of available
-        keys, refer to the `newtonnet.utils.ase_interface.MLAseCalculator` calculator.
-
-        !!! Info "Calculator defaults"
-
-            ```python
-            {
-                "model_path": SETTINGS.NEWTONNET_MODEL_PATH,
-                "settings_path": SETTINGS.NEWTONNET_CONFIG_PATH,
-            }
-            ```
-    opt_swaps
+    opt_params
         Dictionary of custom kwargs for the optimization process. Set a value
         to `None` to remove a pre-existing key entirely. For a list of available
-        keys, refer to [quacc.runners.calc.run_ase_opt][].
+        keys, refer to [quacc.runners.ase.run_opt][].
 
         !!! Info "Optimizer defaults"
 
@@ -208,8 +191,19 @@ def irc_job(
                 },
             }
             ```
-    copy_files
-        Files to copy to the runtime directory.
+    **kwargs
+        Custom kwargs for the NewtonNet calculator. Set a value to
+        `None` to remove a pre-existing key entirely. For a list of available
+        keys, refer to the `newtonnet.utils.ase_interface.MLAseCalculator` calculator.
+
+        !!! Info "Calculator defaults"
+
+            ```python
+            {
+                "model_path": SETTINGS.NEWTONNET_MODEL_PATH,
+                "settings_path": SETTINGS.NEWTONNET_CONFIG_PATH,
+            }
+            ```
 
     Returns
     -------
@@ -238,15 +232,15 @@ def irc_job(
         },
     }
 
-    flags = merge_dicts(defaults, calc_swaps)
-    opt_flags = merge_dicts(opt_defaults, opt_swaps)
+    flags = merge_dicts(defaults, kwargs)
+    opt_flags = merge_dicts(opt_defaults, opt_params)
 
     # Define calculator
     atoms.calc = NewtonNet(**flags)
 
     # Run IRC
     SETTINGS.CHECK_CONVERGENCE = False
-    dyn = run_ase_opt(atoms, copy_files=copy_files, **opt_flags)
+    dyn = run_opt(atoms, **opt_flags)
     opt_irc_summary = _add_stdev_and_hess(
         summarize_opt_run(
             dyn, additional_fields={"name": f"NewtonNet IRC: {direction}"}
@@ -275,7 +269,6 @@ def quasi_irc_job(
     irc_job_kwargs: dict[str, Any] | None = None,
     relax_job_kwargs: dict[str, Any] | None = None,
     freq_job_kwargs: dict[str, Any] | None = None,
-    copy_files: list[str] | None = None,
 ) -> QuasiIRCSchema:
     """
     Perform a quasi-IRC job using the given atoms object. The initial IRC job by default
@@ -295,8 +288,6 @@ def quasi_irc_job(
         Keyword arguments to use for the [quacc.recipes.newtonnet.core.relax_job][]
     freq_job_kwargs
         Keyword arguments to use for the [quacc.recipes.newtonnet.ts.freq_job][]
-    copy_files
-        Files to copy to the runtime directory.
 
     Returns
     -------
@@ -307,7 +298,7 @@ def quasi_irc_job(
     relax_job_kwargs = relax_job_kwargs or {}
     freq_job_kwargs = freq_job_kwargs or {}
 
-    irc_job_defaults = {"calc_swaps": {"max_steps": 5}}
+    irc_job_defaults = {"max_steps": 5}
     irc_job_kwargs = merge_dicts(irc_job_defaults, irc_job_kwargs)
 
     # Run IRC
@@ -315,7 +306,6 @@ def quasi_irc_job(
         atoms,
         direction=direction,
         run_freq=False,
-        copy_files=copy_files,
         **irc_job_kwargs,
     )