+++ src/quacc/recipes/qchem/core.py
@@ -8,7 +8,7 @@ from ase.optimize import FIRE
 
 from quacc import SETTINGS, job
 from quacc.calculators.qchem import QChem
-from quacc.runners.calc import run_ase_calc, run_ase_opt
+from quacc.runners.ase import run_calc, run_opt
 from quacc.schemas.ase import summarize_opt_run, summarize_run
 from quacc.utils.dicts import merge_dicts, remove_dict_nones
 
@@ -30,8 +30,8 @@ if TYPE_CHECKING:
 @job
 def static_job(
     atoms: Atoms,
-    charge: int,
-    spin_multiplicity: int,
+    charge: int = 0,
+    spin_multiplicity: int = 1,
     method: str = "wb97mv",
     basis: str = "def2-tzvpd",
     scf_algorithm: str = "diis",
@@ -135,8 +135,8 @@ def static_job(
 @job
 def internal_relax_job(
     atoms: Atoms,
-    charge: int,
-    spin_multiplicity: int,
+    charge: int = 0,
+    spin_multiplicity: int = 1,
     method: str = "wb97mv",
     basis: str = "def2-svpd",
     scf_algorithm: str = "diis",
@@ -240,8 +240,8 @@ def internal_relax_job(
 @job
 def freq_job(
     atoms: Atoms,
-    charge: int,
-    spin_multiplicity: int,
+    charge: int = 0,
+    spin_multiplicity: int = 1,
     method: str = "wb97mv",
     basis: str = "def2-svpd",
     scf_algorithm: str = "diis",
@@ -334,8 +334,8 @@ def freq_job(
     }
     return _base_job(
         atoms,
-        charge,
-        spin_multiplicity,
+        charge=charge,
+        spin_multiplicity=spin_multiplicity,
         defaults=defaults,
         copy_files=copy_files,
         additional_fields={"name": "Q-Chem Frequency"},
@@ -345,8 +345,8 @@ def freq_job(
 @job
 def relax_job(
     atoms: Atoms,
-    charge: int,
-    spin_multiplicity: int,
+    charge: int = 0,
+    spin_multiplicity: int = 1,
     method: str = "wb97mv",
     basis: str = "def2-svpd",
     scf_algorithm: str = "diis",
@@ -354,7 +354,7 @@ def relax_job(
     smd_solvent: str | None = None,
     n_cores: int | None = None,
     overwrite_inputs: dict[str, Any] | None = None,
-    opt_swaps: dict[str, Any] | None = None,
+    opt_params: dict[str, Any] | None = None,
     copy_files: list[str] | None = None,
 ) -> OptSchema:
     """
@@ -419,8 +419,8 @@ def relax_job(
         Dictionary passed to `pymatgen.io.qchem.QChemDictSet` which can modify
         default values set therein as well as set additional Q-Chem parameters.
         See QChemDictSet documentation for more details.
-    opt_swaps
-        Dictionary of custom kwargs for [quacc.runners.calc.run_ase_opt][]
+    opt_params
+        Dictionary of custom kwargs for [quacc.runners.ase.run_opt][]
     copy_files
         Files to copy to the runtime directory.
 
@@ -453,11 +453,11 @@ def relax_job(
 
     return _base_opt_job(
         atoms,
-        charge,
-        spin_multiplicity,
+        charge=charge,
+        spin_multiplicity=spin_multiplicity,
         qchem_defaults=qchem_defaults,
         opt_defaults=opt_defaults,
-        opt_swaps=opt_swaps,
+        opt_params=opt_params,
         additional_fields={"name": "Q-Chem Optimization"},
         copy_files=copy_files,
     )
@@ -465,8 +465,8 @@ def relax_job(
 
 def _base_job(
     atoms: Atoms,
-    charge: int,
-    spin_multiplicity: int,
+    charge: int = 0,
+    spin_multiplicity: int = 1,
     defaults: dict[str, Any] | None = None,
     additional_fields: dict[str, Any] | None = None,
     copy_files: list[str] | None = None,
@@ -499,7 +499,7 @@ def _base_job(
     qchem_flags = remove_dict_nones(defaults)
 
     atoms.calc = QChem(atoms, **qchem_flags)
-    final_atoms = run_ase_calc(atoms, copy_files=copy_files)
+    final_atoms = run_calc(atoms, copy_files=copy_files)
 
     return summarize_run(
         final_atoms,
@@ -511,11 +511,11 @@ def _base_job(
 
 def _base_opt_job(
     atoms: Atoms,
-    charge: int,
-    spin_multiplicity: int,
+    charge: int = 0,
+    spin_multiplicity: int = 1,
     qchem_defaults: dict[str, Any] | None = None,
     opt_defaults: dict[str, Any] | None = None,
-    opt_swaps: dict[str, Any] | None = None,
+    opt_params: dict[str, Any] | None = None,
     additional_fields: dict[str, Any] | None = None,
     copy_files: list[str] | None = None,
 ) -> OptSchema:
@@ -534,8 +534,8 @@ def _base_opt_job(
         Default arguments for the Q-Chem calculator.
     opt_defaults
         Default arguments for the ASE optimizer.
-    opt_swaps
-        Dictionary of custom kwargs for [quacc.runners.calc.run_ase_opt][]
+    opt_params
+        Dictionary of custom kwargs for [quacc.runners.ase.run_opt][]
     copy_files
         Files to copy to the runtime directory.
 
@@ -548,10 +548,10 @@ def _base_opt_job(
     #   - passing initial Hessian?
 
     qchem_flags = remove_dict_nones(qchem_defaults)
-    opt_flags = merge_dicts(opt_defaults, opt_swaps)
+    opt_flags = merge_dicts(opt_defaults, opt_params)
 
     atoms.calc = QChem(atoms, **qchem_flags)
-    dyn = run_ase_opt(atoms, copy_files=copy_files, **opt_flags)
+    dyn = run_opt(atoms, copy_files=copy_files, **opt_flags)
 
     return summarize_opt_run(
         dyn,