+++ tests/core/recipes/emt_recipes/test_emt_recipes.py
@@ -18,7 +18,7 @@ def test_static_job(tmpdir):
     assert output["parameters"]["asap_cutoff"] is False
     assert output["results"]["energy"] == pytest.approx(0.07001766638245854)
 
-    output = static_job(atoms, calc_swaps={"asap_cutoff": True})
+    output = static_job(atoms, asap_cutoff=True)
     assert output["nsites"] == len(atoms)
     assert output["parameters"]["asap_cutoff"] is True
     assert output["results"]["energy"] == pytest.approx(0.11074520235398744)
@@ -63,8 +63,8 @@ def test_relax_job(tmpdir):
     atoms[0].position += [0.1, 0.1, 0.1]
     output = relax_job(
         atoms,
-        opt_swaps={"fmax": 0.03},
-        calc_swaps={"asap_cutoff": True},
+        opt_params={"fmax": 0.03},
+        asap_cutoff=True,
     )
     assert output["nsites"] == len(atoms)
     assert output["parameters"]["asap_cutoff"] is True
@@ -77,8 +77,8 @@ def test_relax_job(tmpdir):
     atoms.set_constraint(c)
     output = relax_job(
         atoms,
-        opt_swaps={"fmax": 0.03},
-        calc_swaps={"asap_cutoff": True},
+        opt_params={"fmax": 0.03},
+        asap_cutoff=True,
     )
     assert output["nsites"] == len(atoms)
     assert output["parameters"]["asap_cutoff"] is True
@@ -102,10 +102,7 @@ def test_slab_dynamic_jobs(tmpdir):
     outputs = bulk_to_slabs_flow(
         atoms,
         run_static=False,
-        slab_relax_kwargs={
-            "opt_swaps": {"fmax": 1.0},
-            "calc_swaps": {"asap_cutoff": True},
-        },
+        slab_relax_kwargs={"opt_params": {"fmax": 1.0}, "asap_cutoff": True},
     )
     assert len(outputs) == 4
     assert outputs[0]["nsites"] == 80