+++ tests/core/schemas/test_prep.py
@@ -8,7 +8,7 @@ from ase.io import read
 
 from quacc.atoms.core import get_atoms_id
 from quacc.calculators.vasp import Vasp
-from quacc.runners.prep import prep_next_run
+from quacc.schemas.prep import prep_next_run
 
 
 @pytest.fixture()